자유게시판 페이징 기능 구현

+++ rental-project/src/main/java/com/rentalproject/controller/FreeBoardController.java
@@ -21,6 +21,7 @@ import com.rentalproject.common.Util;
 import com.rentalproject.dto.FreeBoardAttachDto;
 import com.rentalproject.dto.FreeBoardDto;
 import com.rentalproject.service.FreeBoardService;
+import com.rentalproject.ui.ThePager;
 import com.rentalproject.view.DownloadView;
 
 @Controller
@@ -32,11 +33,24 @@ public class FreeBoardController {
 	
 	// 자유게시글 리스트 화면 불러오기 ( 전체 게시글 불러오기 )
 	@GetMapping(path= {"/freeboardlist"})
-	public String list(Model model) { 
+	public String list(@RequestParam(defaultValue = "1") int pageNo, Model model) { 
 		
-		List<FreeBoardDto> freeBoardList = freeBoardService.listFreeBoard();
+		//List<FreeBoardDto> freeBoardList = freeBoardService.listFreeBoard();  // 전체 게시물 조회
+		
+		int pageSize = 10;
+		int pagerSize = 5;
+		String linkUrl = "freeboardlist";
+		int dataCount = freeBoardService.getFreeBoardCount();
+		
+		int from = (pageNo -1) * pageSize;
+		List<FreeBoardDto> freeBoardList = freeBoardService.listFreeBoardByPage(from, pageSize);
+		
+		ThePager pager = new ThePager(dataCount, pageNo, pageSize, pagerSize, linkUrl);
+		
 		 
 		model.addAttribute("freeBoardList", freeBoardList);
+		model.addAttribute("pager", pager);
+		model.addAttribute("pageNo", pageNo);
 		
 		return "freeboard/freeboardlist";
 	}
@@ -53,10 +67,7 @@ public class FreeBoardController {
 	// 자유게시글 등록하기
 	@PostMapping(path= {"/freeboardwrite"})
 	
-		public String writeFreeBoard(FreeBoardDto freeboard, MultipartFile attach, HttpServletRequest req) throws Exception {
-		
-		System.out.println("제목:" + freeboard.getFreeBoardTitle());
-		System.out.println("내용:" + freeboard.getFreeBoardContent());
+		public String writeFreeBoard(FreeBoardDto freeboard, MultipartFile attach, HttpServletRequest req) throws Exception { 
 		
 		// 파일업로드 처리
 		String uploadAttachFile = req.getServletContext().getRealPath("/resources/upload/");
@@ -96,9 +107,11 @@ public class FreeBoardController {
 
 	// 자유게시글 클릭 후 상세보기
 	@GetMapping(path = {"/freeboarddetail"})
-	public String detail(@RequestParam(defaultValue = "-1") int freeBoardNo, Model model) {
+	public String detail(@RequestParam(defaultValue = "-1") int freeBoardNo,
+						 @RequestParam(defaultValue = "-1") int pageNo,
+						 Model model) {
 		
-		if(freeBoardNo == -1) {  // 주소창에 detail로 바로 접근하지 못하게 함 
+		if(freeBoardNo == -1 || pageNo == -1) {  // 주소창에 detail로 바로 접근하지 못하게 함 
 			return "redirect:freeboardlist";
 		}
 		
@@ -109,6 +122,9 @@ public class FreeBoardController {
 		}
 		
 		model.addAttribute("freeBoard", freeboard);
+		model.addAttribute("pageNo", pageNo);
+		
+		freeBoardService.updateFreeBoardviewCount(freeBoardNo);
 		
 		return "freeboard/freeboarddetail";
 		
@@ -131,25 +147,34 @@ public class FreeBoardController {
 		// 자유게시글 수정하기 ( 자유게시글 상세보기 내용 불러오기 )
 		@GetMapping(path = {"/freeboardedit"})
 		public String showFreeBoardEditForm(@RequestParam(defaultValue = "-1") int freeBoardNo, 
+											@RequestParam(defaultValue = "-1") int pageNo,
 											Model model) {
-		
+			
+			if (freeBoardNo == -1 || pageNo == -1) {
+				return "redirect:freeboardlist";
+			}
+			
 		FreeBoardDto freeboard = freeBoardService.findFreeBoardByFreeBoardNo(freeBoardNo);
 		
+			if (freeboard == null) {
+				return "redirect:freeboardlist";
+			}
 		
-		if(freeboard == null) { // 조회된 글이 없을때 리스트로 
-			
-			return "redirect:freeboardlist";
-		}
-		 
 		model.addAttribute("freeBoard", freeboard);
+		model.addAttribute("pageNo", pageNo);
 		
 		return "freeboard/freeboardedit";
 	}
 	
 	// 자유게시글 수정하기 ( 수정한 글 등록하기 )
 		@PostMapping(path = {"/freeboardedit"})
-		public String freeBoardEdit(FreeBoardDto freeboard, MultipartFile attach, HttpServletRequest req) {
+		public String freeBoardEdit(FreeBoardDto freeboard, MultipartFile attach, HttpServletRequest req,
+									@RequestParam(defaultValue = "-1") int pageNo) {
 		
+			if (pageNo < 1) {
+				return "redirect:freeboardlist";
+			}
+			
 		String uploadAttachFile = req.getServletContext().getRealPath("/resources/upload/");
 		ArrayList<FreeBoardAttachDto> freeBoardAttachList = handleUploadFile(attach, uploadAttachFile);
 		freeboard.setFreeBoardAttachList(freeBoardAttachList);
@@ -157,35 +182,20 @@ public class FreeBoardController {
 		// update 처리하기
 		freeBoardService.editFreeBoard(freeboard);
 		
-		return "redirect:freeboardlist";
+		return String.format("redirect:freeboarddetail?freeBoardNo=%d&pageNo=%d", freeboard.getFreeBoardNo(), pageNo);
 	}
 	
 	// 자유게시글 삭제하기
 	@GetMapping(path = {"/freeboarddelete/{freeBoardNo}" })
-	public String freeBoardDelete(@PathVariable("freeBoardNo") int freeBoardNo ) {
+	public String freeBoardDelete(@PathVariable("freeBoardNo") int freeBoardNo,
+								  @RequestParam(defaultValue = "-1") int pageNo) {
 		
+		if (pageNo == -1) {
+			return "redirect:/freeboard/freeboardlist";
+		}
 		freeBoardService.deleteFreeBoard(freeBoardNo);
-		
-		return "redirect:/freeboard/freeboardlist";	
+		return String.format("redirect:/freeboard/freeboardlist?pageNo=%d", pageNo);
 	}
-	
-	
-	
-	  // 자유게시판 조회수 증가
-		/*
-		 * @GetMapping(path = {"/freeboardlist"}) public String
-		 * FreeBoardviewCountUpdate(@RequestParam(name="freeBoardNo") int freeBoardNo,
-		 * Model model) {
-		 * 
-		 * freeBoardService.updateFreeBoardviewCount(freeBoardNo);
-		 * 
-		 * model.addAttribute("freeBoardNo", freeBoardNo);
-		 * 
-		 * return "freeboard/freeboardlist"; }
-		 */
-	 
-
-	
 		
 }
